Move sub-classes of TreeNode class to separate file
