+++ src/System.Windows.Forms/src/System/Windows/Forms/TreeNode.TreeNodeImageIndexer.ImageListType.cs
@@ -1,0 +1,18 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    public partial class TreeNode
+    {
+        internal partial class TreeNodeImageIndexer
+        {
+            public enum ImageListType
+            {
+                Default,
+                State
+            }
+        }
+    }
+}
